smartexam-compiler/requirements.txt
flask>=3.0.0
werkzeug>=3.0.0
pymupdf>=1.23.0
pytesseract>=0.3.10
pillow>=10.0.0
pdf2image>=1.16.3
reportlab>=4.0.0
numpy>=1.24.0
matplotlib>=3.7.0
google-generativeai>=0.3.0
